Cosmetic changes
Several cosmetic changes on Tutorial examples.
The most important is to try to enhance the model visibility when shown in a small space. This is especially useful when you want to see a diagram window in the plotting view in OpernModelica.
In case this approach is accepted, I can make similar changes also for the other examples within the tutorial folder, and even more.