LoadImpedancePQ rewritten without extend the BusBase model
The implemented propagation of the UStart and UPhaseStart values from
the bus to the connected components doesn't allow to use the BusBase
model to build components different from Buses.
This because the BusBase interface has output connectors that propagate
said values while non-Bus components (like Loads) shall have in the
interface the corrispondent input connectors that receive said values
from the bus.