terminalACPF removed if EPF is not computed, fixed wrong condition
NOTE:
terminalACPF has been removed if the embedded PF is not computed because
in this case the variables vPF and iPF are forced to Complex(0) in each
xxPortAC interface with the result that each xxPortAC interface
connected to the same PF connection set (generated by the terminalACPF)
generates  four eaquations (vPF=Complex(0) and iPF=Complex(0)) that
involve the same pair of complex variables in the connection set, thus
generating more equations than necessary. The result is that the overall
model becomes unbalanced.

The problem can be easly solved by removing the PF connection sets (that
is by removing the terminalACPF) if the embedded PF is not computed.